requirements.txt
python-chess>=1.9.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.2.0
xgboost>=1.7.0
streamlit>=1.22.0
stockfish>=3.28.0
python-dotenv>=0.21.0